Remove old SideBar
