+++ src/renderer/components/TimelineSpace/Contents/DirectMessages.vue
@@ -13,23 +13,17 @@
             :server="account.server"
             v-on:update="updateToot"
             v-on:delete="deleteToot"
-            @focusRight="focusSidebar"
             @selectToot="focusToot(item)"
           >
           </toot>
         </DynamicScrollerItem>
       </template>
     </DynamicScroller>
-    <div :class="openSideBar ? 'upper-with-side-bar' : 'upper'" v-show="!heading">
-      <el-button type="primary" @click="upper" circle>
-        <font-awesome-icon icon="angle-up" class="upper-icon" />
-      </el-button>
-    </div>
   </div>
 </template>
 
 <script lang="ts">
-import { defineComponent, ref, computed, onMounted, onBeforeUpdate, onBeforeUnmount, watch, reactive } from 'vue'
+import { defineComponent, ref, computed, onMounted, onBeforeUpdate, watch, reactive } from 'vue'
 import { logicAnd } from '@vueuse/math'
 import { useMagicKeys, whenever } from '@vueuse/core'
 import { useStore } from '@/store'
@@ -38,7 +32,6 @@ import { useRoute } from 'vue-router'
 import { ElMessage } from 'element-plus'
 import { Entity } from 'megalodon'
 import Toot from '@/components/organisms/Toot.vue'
-import { EventEmitter } from '@/components/event'
 import { ACTION_TYPES, MUTATION_TYPES } from '@/store/TimelineSpace/Contents/DirectMessages'
 import { MUTATION_TYPES as SIDE_MENU_MUTATION } from '@/store/TimelineSpace/SideMenu'
 import { LocalAccount } from '~/src/types/localAccount'
@@ -68,7 +61,6 @@ export default defineComponent({
     })
 
     const timeline = computed(() => store.state.TimelineSpace.Contents.DirectMessages.timeline[id.value])
-    const openSideBar = computed(() => store.state.TimelineSpace.Contents.SideBar.openSideBar)
     const modalOpened = computed<boolean>(() => store.getters[`TimelineSpace/Modals/modalOpened`])
     const currentFocusedIndex = computed(() => timeline.value.findIndex(toot => focusedId.value === toot.uri + toot.id))
     const shortcutEnabled = computed(() => !modalOpened.value)
@@ -85,9 +77,6 @@ export default defineComponent({
       if (store.state.TimelineSpace.SideMenu.unreadDirectMessagesTimeline && heading.value) {
         store.commit(`TimelineSpace/SideMenu/${SIDE_MENU_MUTATION.CHANGE_UNREAD_DIRECT_MESSAGES_TIMELINE}`, false)
       }
-    })
-    onBeforeUnmount(() => {
-      EventEmitter.off('focus-timeline')
     })
     watch(focusedId, (newVal, _oldVal) => {
       if (newVal && heading.value) {
@@ -169,9 +158,6 @@ export default defineComponent({
     const focusToot = (message: Entity.Status) => {
       focusedId.value = message.uri + message.id
     }
-    const focusSidebar = () => {
-      EventEmitter.emit('focus-sidebar')
-    }
 
     return {
       timeline,
@@ -180,9 +166,7 @@ export default defineComponent({
       modalOpened,
       updateToot,
       deleteToot,
-      focusSidebar,
       focusToot,
-      openSideBar,
       heading,
       upper,
       account
@@ -207,24 +191,6 @@ export default defineComponent({
 
   .loading-card:empty {
     height: 0;
-  }
-
-  .upper {
-    position: fixed;
-    bottom: 20px;
-    right: 20px;
-    transition: all 0.5s;
-  }
-
-  .upper-with-side-bar {
-    position: fixed;
-    bottom: 20px;
-    right: calc(20px + var(--current-sidebar-width));
-    transition: all 0.5s;
-  }
-
-  .upper-icon {
-    padding: 3px;
   }
 }
 </style>
